Feature: Packed Mode Transmission Working. Interface and Firmware: Bugs Fixed. Please Read the Description.
Changes in Interface Code:

- New Class respossible by Unpacking Samples Added (Unpacker).
- New Command to change the Transmission Mode was created.
- Checkbox "Transmission Mode" connected to the command that change the Transmission Mode.
- Method "receive()" was changed to Support the new Transmission Mode.
- Bug related to transmission parameters change was removed.
- New way of verify if the Log need to be Saved After a Stop Capture Comand was implemented.
- Method  to Sync the Board was Updated.

Changes in Firmware Code:

- New Command to change the Transmission Mode was created.
- Validation before aply the new command was implemented.
- Commands that change the Transmission Mode were conected to the transmission code.
- Corrections in the Recalculation of Number of Instats captured were maded in four of the "Set" functions,
in the comunication module.
- Indexes in Loops related to the Packed Transmission Mode were revised and corrected.
- Bug related to the use of bits per sample different to 12 in Unpack Transmission Mode was Fixed.

IMPORTANT!!!

- Please Update the Firmware Before use that Version of Interface.
Bugs in Packed Transmission Mode were removed. Burn EMG2.bin file in the Project Folder into the Tiva Flash.

- A Performance Limitation caused by Processing Power of the PC running the Interface was noted
when Both Transmission Mode were Used when higher Sample Rate were used.

- Note that the Packed Transmission it is a bit heavier than the Unpacked Transmission.

- If you experience Performance Issues after a Change in the Capture Parameters while using Packed Transmission,
Please use a different Packet Size to the rebalance the time of Capture and Process the Samples.

- There are a Optimum Packet Size to balance the time of processing of Capture and Process of the Samples.
A auto Packet Size would be great to choose Sizes after a change in the capture parameters.

- The Calculation of Sine Samples when using Function Generator takes a good amount processing.
Please use Square Wave or Sawtooth Wave for higher Sample Rate.
A new version of Sine Calculation will be implemented in the future requiring much less amount processing.

+++ software/firmware/Project Folder/EMG2/function_gen.c
@@ -69,7 +69,7 @@ uint16_t function_gen(uint8_t wave_form_, uint32_t timestamp_, uint32_t func_gen
 
 uint16_t sin_wave_gen(float pecent_of_period_)     {                                       // The argument to Sine function it is the sample number inside of a period of the function.
 
-    double sin_argument = pecent_of_period_ * 2 * M_PI;                                    //  The angle of Sine Function will be a percentage of 2 * Pi.
+    float sin_argument = pecent_of_period_ * 2 * M_PI;                                    //  The angle of Sine Function will be a percentage of 2 * Pi.
                                                                                            //  See Note 1.
 
     return (uint16_t) round( Highest_12b_Num * ( (sin( sin_argument ) + 1) / 2) ) ;        //  Shifts the Sine value by 1 than divide the result by 2  to make the the value
